Rework adapter tests namespaces to be consistent
