+++ infection.json
@@ -3,14 +3,15 @@
     "directories": [
         "src/core/etl/src",
         "src/adapter/etl-adapter-avro/src",
-        "src/adapter/etl-adapter-csv/src",
         "src/adapter/etl-adapter-chartjs/src",
+        "src/adapter/etl-adapter-csv/src",
         "src/adapter/etl-adapter-doctrine/src",
         "src/adapter/etl-adapter-elasticsearch/src",
         "src/adapter/etl-adapter-google-sheet/src",
         "src/adapter/etl-adapter-http/src",
         "src/adapter/etl-adapter-json/src",
         "src/adapter/etl-adapter-logger/src",
+        "src/adapter/etl-adapter-meilisearch/src",
         "src/adapter/etl-adapter-parquet/src",
         "src/adapter/etl-adapter-text/src",
         "src/adapter/etl-adapter-xml/src",
