+++ src/adapter/etl-adapter-avro/tests/Flow/ETL/Adapter/Avro/Tests/Fixtures/data.avro
