+++ src/adapter/etl-adapter-http/tests/Flow/ETL/Adapter/HTTP/Tests/Fixtures/norberttech.json
